+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [3.57.1] - 2020-05-06
+### Added
+- `TotalProductsFlexible` and `OrderByFlexible` to the site-editor.
 
 ### Fixed